Extract folder tree hit target
